refactor(plugins): inline map and socialSharing config modals into plugin dirs
PanelSettingsModalMapViewer and PanelSettingsModalSocialMedia were thin
wrappers sitting in app/components/modals/ while the actual plugin wrapper
lived in plugins/. Collapsed both into their plugin configModal files
directly and deleted the intermediate components.
